+++ packages/scaffold-react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @web3modal/scaffold-react
 
+## 5.0.6
+
+### Patch Changes
+
+- fix: Social Login illegal invocation issue. Wagmi tests
+
+- Updated dependencies []:
+  - @web3modal/scaffold@5.0.6
+
 ## 5.0.5
 
 ### Patch Changes